refactor: review applied
