--- src/entities/chat-item/index.ts
@@ -1,1 +1,0 @@
-export { default as ChatItem } from './ui/chat-item.component';
