+++ src/widgets/partyroom-chat-panel/index.ts
@@ -1,0 +1,1 @@
+export { default as PartyroomChatPanel } from './ui/partyroom-chat-panel.component';
