+++ src/widgets/partyroom-chat-panel/ui/chat-item/authority-headset.component.tsx
