+++ src/widgets/partyroom-chat-panel/ui/chat-item/index.ts
@@ -1,0 +1,1 @@
+export { default as ChatItem } from './chat-item.component';
