--- src/widgets/partyroom-tab-panel/config/partyroom-chat-tab-panel-size.config.ts
@@ -1,9 +1,0 @@
-/**
- * 채팅 탭 패널의 시작 위치
- */
-export const CHAT_TAB_PANEL_START_POSITION = 130;
-
-/**
- * 채팅 탭 패널의 끝 위치
- */
-export const CHAT_TAB_PANEL_END_POSITION = 96;
