--- src/widgets/partyroom-tab-panel/index.ts
@@ -1,1 +1,0 @@
-export { default as PartyroomChatTabPanel } from './ui/partyroom-chat-tab-panel.component';
